PHD-69 Train binary classifier (#8)
* Decent baseline binary mnist model

* Add evaluate model performance script

* flake8

* Remove commented lines

+++ xai/data_handlers/mnist.py
@@ -108,7 +108,7 @@ def load_mnist_binary_dataset(
             torchvision.transforms.ToTensor(),
             torchvision.transforms.Normalize(*DEFAULT_MNIST_NORMALIZATION),
         ]),
-        target_transform=lambda x: torch.tensor((x == target_digit), dtype=torch.float32)
+        target_transform=lambda x: torch.tensor((x == target_digit), dtype=torch.long)
     )
 
     if digits:
